Renamed hash to duplicates and created new file hashes
